+++ starling_sim/utils/paths.py
@@ -68,9 +68,15 @@ Structure of the data folder
 
 The path to the data repository and the names of the folders can be changed.
 
+The path to the data repository can be changed using the --data-folder option of main.py
+
+.. code-block:: bash
+
+    python3 main.py path/to/parameters.json --data-folder path/to/data_folder
+
 However, the structure of the data repository must remain the same. This is ensured by the functions
-listed below, that build the paths to the different folders by concatenating the folder names
-in the correct order.
+declared in starling_sim.utils.paths.py, that build the paths to the different folders by concatenating
+the folder names in the correct order.
 
 ***************************
 Structure of models package